plugins: migrate single-surface modules to the `props` mapper
Registers each container as itself and maps the surface props onto its own props,
so the unpacking of `data` is checked against the surface's filter instead of being
restated in an inline arrow. Every capability module here is now JSX-free and
renamed to `.ts`.

Where the inline arrow called hooks it cannot become a mapper, so it is extracted
as a named component beside the capability module (the `useSettingsState` binding
in native/payments/observability/excalidraw/crx, and the active-workspace lookup
in native-filesystem).
