+++ packages/plugins/plugin-brain/src/capabilities/react-surface.ts
@@ -1,0 +1,24 @@
+//
+// Copyright 2026 DXOS.org
+//
+
+import * as Effect from 'effect/Effect';
+
+import { Capabilities, Capability } from '@dxos/app-framework';
+import { Surface } from '@dxos/app-framework/ui';
+
+import { FactsCompanion } from '#containers';
+import { BrainSurface } from '#types';
+
+/** React surfaces contributed by plugin-brain — the per-space facts panel. */
+export default Capability.makeModule(() =>
+  Effect.succeed(
+    Capability.contributes(Capabilities.ReactSurface, [
+      Surface.create({
+        id: 'brain.facts',
+        filter: Surface.makeFilter(BrainSurface.Facts),
+        component: FactsCompanion,
+      }),
+    ]),
+  ),
+);
