+++ packages/plugins/plugin-chess-com/src/capabilities/react-surface.ts
@@ -3,27 +3,25 @@
 //
 
 import * as Effect from 'effect/Effect';
-import React from 'react';
 
 import { Capabilities, Capability } from '@dxos/app-framework';
 import { Surface } from '@dxos/app-framework/ui';
 import { AppSurface } from '@dxos/app-toolkit/ui';
 
-import { TerraArticle } from '#containers';
-import { Terra } from '#types';
+import { ChessGameArticle } from '#containers';
+import { ChessComAccount } from '#types';
 
 export default Capability.makeModule(() =>
   Effect.succeed(
     Capability.contributes(Capabilities.ReactSurface, [
       Surface.create({
-        id: 'terra',
+        id: 'chessGameArticle',
         filter: AppSurface.oneOf(
-          AppSurface.object(AppSurface.Article, Terra.Terra),
-          AppSurface.object(AppSurface.Section, Terra.Terra),
+          AppSurface.object(AppSurface.Article, ChessComAccount.Account),
+          AppSurface.object(AppSurface.Section, ChessComAccount.Account),
         ),
-        component: ({ data, role }) => {
-          return <TerraArticle role={role} subject={data.subject} attendableId={data.attendableId} />;
-        },
+        component: ChessGameArticle,
+        props: ({ role, data: { subject, attendableId } }) => ({ role, subject, attendableId }),
       }),
     ]),
   ),
