+++ packages/plugins/plugin-crx/src/capabilities/react-surface.ts
@@ -3,15 +3,14 @@
 //
 
 import * as Effect from 'effect/Effect';
-import React from 'react';
 
 import { Capabilities, Capability } from '@dxos/app-framework';
-import { Surface, useSettingsState } from '@dxos/app-framework/ui';
+import { Surface } from '@dxos/app-framework/ui';
 import { AppSurface } from '@dxos/app-toolkit/ui';
 
-import { PaymentsSettings } from '#containers';
 import { meta } from '#meta';
-import { type Settings } from '#types';
+
+import { CrxSettingsSurface } from './CrxSettingsSurface';
 
 export default Capability.makeModule(() =>
   Effect.succeed(
@@ -19,10 +18,8 @@ export default Capability.makeModule(() =>
       Surface.create({
         id: 'pluginSettings',
         filter: AppSurface.settings(AppSurface.Article, meta.profile.key),
-        component: ({ data: { subject } }) => {
-          const { settings, updateSettings } = useSettingsState<Settings.Settings>(subject.atom);
-          return <PaymentsSettings settings={settings} onSettingsChange={updateSettings} />;
-        },
+        component: CrxSettingsSurface,
+        props: ({ data: { subject } }) => ({ subject }),
       }),
     ]),
   ),
