+++ packages/plugins/plugin-doctor/src/capabilities/react-surface.ts
@@ -3,23 +3,25 @@
 //
 
 import * as Effect from 'effect/Effect';
-import React from 'react';
 
 import { Capabilities, Capability } from '@dxos/app-framework';
 import { Surface } from '@dxos/app-framework/ui';
 import { AppSurface } from '@dxos/app-toolkit/ui';
-import { Project } from '@dxos/compute';
 
-import { ProjectArticle } from '#containers';
+import { DiagnosticsPanel } from '#containers';
 
-/** React surfaces contributed by plugin-projects — the Project detail article. */
+import { DIAGNOSTICS_DECK_COMPANION_ID } from './app-graph-builder';
+
 export default Capability.makeModule(() =>
   Effect.succeed(
     Capability.contributes(Capabilities.ReactSurface, [
       Surface.create({
-        id: 'project.article',
-        filter: AppSurface.object(AppSurface.Article, Project.Project),
-        component: ({ data, role }) => <ProjectArticle role={role} {...data} />,
+        id: 'doctorDiagnostics',
+        filter: AppSurface.literal(
+          AppSurface.deckCompanion(DIAGNOSTICS_DECK_COMPANION_ID),
+          DIAGNOSTICS_DECK_COMPANION_ID,
+        ),
+        component: DiagnosticsPanel,
       }),
     ]),
   ),
