--- packages/plugins/plugin-excalidraw/src/capabilities/react-surface.tsx
@@ -1,29 +1,0 @@
-//
-// Copyright 2025 DXOS.org
-//
-
-import * as Effect from 'effect/Effect';
-import React from 'react';
-
-import { Capabilities, Capability } from '@dxos/app-framework';
-import { Surface, useSettingsState } from '@dxos/app-framework/ui';
-import { AppSurface } from '@dxos/app-toolkit/ui';
-
-import { ExcalidrawSettings } from '#containers';
-import { meta } from '#meta';
-import { type Settings } from '#types';
-
-export default Capability.makeModule(() =>
-  Effect.succeed(
-    Capability.contributes(Capabilities.ReactSurface, [
-      Surface.create({
-        id: 'plugin-settings',
-        filter: AppSurface.settings(AppSurface.Article, meta.profile.key),
-        component: ({ data: { subject } }) => {
-          const { settings, updateSettings } = useSettingsState<Settings.Settings>(subject.atom);
-          return <ExcalidrawSettings settings={settings} onSettingsChange={updateSettings} />;
-        },
-      }),
-    ]),
-  ),
-);
