+++ packages/plugins/plugin-native-filesystem/src/capabilities/react-surface.ts
@@ -3,15 +3,14 @@
 //
 
 import * as Effect from 'effect/Effect';
-import React from 'react';
 
 import { Capabilities, Capability } from '@dxos/app-framework';
 import { Surface } from '@dxos/app-framework/ui';
 import { AppSurface } from '@dxos/app-toolkit/ui';
 
-import { WorkspaceSettingsContainer } from '#containers';
-import { useActiveFilesystemWorkspace } from '#hooks';
 import { meta } from '#meta';
+
+import { WorkspaceSettingsSurface } from './WorkspaceSettingsSurface';
 
 const GENERAL_TYPE = `${meta.profile.key}.general`;
 
@@ -21,14 +20,7 @@ export default Capability.makeModule(
       Surface.create({
         id: 'workspaceSettings',
         filter: AppSurface.literal(AppSurface.Article, GENERAL_TYPE),
-        component: () => {
-          const workspace = useActiveFilesystemWorkspace();
-          if (!workspace) {
-            return null;
-          }
-
-          return <WorkspaceSettingsContainer workspace={workspace} />;
-        },
+        component: WorkspaceSettingsSurface,
       }),
     ]);
   }),
