+++ packages/plugins/plugin-native/src/capabilities/react-surface.ts
@@ -1,0 +1,26 @@
+//
+// Copyright 2025 DXOS.org
+//
+
+import * as Effect from 'effect/Effect';
+
+import { Capabilities, Capability } from '@dxos/app-framework';
+import { Surface } from '@dxos/app-framework/ui';
+import { AppSurface } from '@dxos/app-toolkit/ui';
+
+import { meta } from '#meta';
+
+import { NativeSettingsSurface } from './NativeSettingsSurface';
+
+export default Capability.makeModule(() =>
+  Effect.succeed(
+    Capability.contributes(Capabilities.ReactSurface, [
+      Surface.create({
+        id: 'pluginSettings',
+        filter: AppSurface.settings(AppSurface.Article, meta.profile.key),
+        component: NativeSettingsSurface,
+        props: ({ data: { subject } }) => ({ subject }),
+      }),
+    ]),
+  ),
+);
