+++ packages/plugins/plugin-observability/src/capabilities/react-surface.ts
@@ -3,26 +3,23 @@
 //
 
 import * as Effect from 'effect/Effect';
-import React from 'react';
 
 import { Capabilities, Capability } from '@dxos/app-framework';
-import { Surface, useSettingsState } from '@dxos/app-framework/ui';
+import { Surface } from '@dxos/app-framework/ui';
 import { AppSurface } from '@dxos/app-toolkit/ui';
 
-import { NativeSettings } from '#containers';
 import { meta } from '#meta';
-import { type Settings } from '#types';
+
+import { ObservabilitySettingsSurface } from './ObservabilitySettingsSurface';
 
 export default Capability.makeModule(() =>
   Effect.succeed(
     Capability.contributes(Capabilities.ReactSurface, [
       Surface.create({
-        id: 'pluginSettings',
+        id: 'root',
         filter: AppSurface.settings(AppSurface.Article, meta.profile.key),
-        component: ({ data: { subject } }) => {
-          const { settings, updateSettings } = useSettingsState<Settings.Settings>(subject.atom);
-          return <NativeSettings settings={settings} onSettingsChange={updateSettings} />;
-        },
+        component: ObservabilitySettingsSurface,
+        props: ({ data: { subject } }) => ({ subject }),
       }),
     ]),
   ),
