+++ packages/plugins/plugin-progress/src/capabilities/react-surface.ts
@@ -3,22 +3,20 @@
 //
 
 import * as Effect from 'effect/Effect';
-import React from 'react';
 
 import { Capabilities, Capability } from '@dxos/app-framework';
 import { Surface } from '@dxos/app-framework/ui';
+import { AppSurface } from '@dxos/app-toolkit/ui';
 
-import { FactsCompanion } from '#containers';
-import { BrainSurface } from '#types';
+import { ProgressStatusIndicator } from '#components';
 
-/** React surfaces contributed by plugin-brain — the per-space facts panel. */
 export default Capability.makeModule(() =>
   Effect.succeed(
     Capability.contributes(Capabilities.ReactSurface, [
       Surface.create({
-        id: 'brain.facts',
-        filter: Surface.makeFilter(BrainSurface.Facts),
-        component: () => <FactsCompanion />,
+        id: 'progressStatusIndicator',
+        filter: Surface.makeFilter(AppSurface.StatusIndicator),
+        component: ProgressStatusIndicator,
       }),
     ]),
   ),
