+++ packages/plugins/plugin-projects/src/capabilities/react-surface.ts
@@ -3,21 +3,23 @@
 //
 
 import * as Effect from 'effect/Effect';
-import React from 'react';
 
 import { Capabilities, Capability } from '@dxos/app-framework';
 import { Surface } from '@dxos/app-framework/ui';
 import { AppSurface } from '@dxos/app-toolkit/ui';
+import { Project } from '@dxos/compute';
 
-import { BeaconStatusIndicator } from '#components';
+import { ProjectArticle } from '#containers';
 
+/** React surfaces contributed by plugin-projects — the Project detail article. */
 export default Capability.makeModule(() =>
   Effect.succeed(
     Capability.contributes(Capabilities.ReactSurface, [
       Surface.create({
-        id: 'beaconStatus',
-        filter: Surface.makeFilter(AppSurface.StatusIndicator),
-        component: () => <BeaconStatusIndicator />,
+        id: 'project.article',
+        filter: AppSurface.object(AppSurface.Article, Project.Project),
+        component: ProjectArticle,
+        props: ({ role, data }) => ({ role, ...data }),
       }),
     ]),
   ),
