+++ packages/plugins/plugin-sequencer/src/capabilities/react-surface.ts
@@ -3,27 +3,25 @@
 //
 
 import * as Effect from 'effect/Effect';
-import React from 'react';
 
 import { Capabilities, Capability } from '@dxos/app-framework';
 import { Surface } from '@dxos/app-framework/ui';
 import { AppSurface } from '@dxos/app-toolkit/ui';
 
-import { ChessGameArticle } from '#containers';
-import { ChessComAccount } from '#types';
+import { ScoreArticle } from '#containers';
+import { Score } from '#types';
 
 export default Capability.makeModule(() =>
   Effect.succeed(
     Capability.contributes(Capabilities.ReactSurface, [
       Surface.create({
-        id: 'chessGameArticle',
+        id: 'article',
         filter: AppSurface.oneOf(
-          AppSurface.object(AppSurface.Article, ChessComAccount.Account),
-          AppSurface.object(AppSurface.Section, ChessComAccount.Account),
+          AppSurface.object(AppSurface.Article, Score.Score),
+          AppSurface.object(AppSurface.Section, Score.Score),
         ),
-        component: ({ data, role }) => (
-          <ChessGameArticle role={role} subject={data.subject} attendableId={data.attendableId} />
-        ),
+        component: ScoreArticle,
+        props: ({ role, data: { subject, attendableId } }) => ({ role, subject, attendableId }),
       }),
     ]),
   ),
