+++ packages/plugins/plugin-stack/src/capabilities/react-surface.ts
@@ -3,25 +3,23 @@
 //
 
 import * as Effect from 'effect/Effect';
-import React from 'react';
 
 import { Capabilities, Capability } from '@dxos/app-framework';
 import { Surface } from '@dxos/app-framework/ui';
 import { AppSurface } from '@dxos/app-toolkit/ui';
+import { Collection } from '@dxos/echo';
 
-import { TemplatePanel } from '#components';
-import { Template } from '#types';
+import { StackArticle } from '#containers';
 
 export default Capability.makeModule(() =>
   Effect.succeed(
     Capability.contributes(
       Capabilities.ReactSurface,
       Surface.create({
-        id: 'root',
-        filter: AppSurface.object(AppSurface.Article, Template.Data),
-        component: ({ data, role }) => (
-          <TemplatePanel role={role} subject={data.subject} attendableId={data.attendableId} />
-        ),
+        id: 'article',
+        filter: AppSurface.object(AppSurface.Article, Collection.Collection),
+        component: StackArticle,
+        props: ({ role, data: { subject, attendableId } }) => ({ role, subject, attendableId }),
       }),
     ),
   ),
