+++ packages/plugins/plugin-template/src/capabilities/react-surface.ts
@@ -3,25 +3,23 @@
 //
 
 import * as Effect from 'effect/Effect';
-import React from 'react';
 
 import { Capabilities, Capability } from '@dxos/app-framework';
 import { Surface } from '@dxos/app-framework/ui';
 import { AppSurface } from '@dxos/app-toolkit/ui';
-import { Collection } from '@dxos/echo';
 
-import { StackArticle } from '#containers';
+import { TemplatePanel } from '#components';
+import { Template } from '#types';
 
 export default Capability.makeModule(() =>
   Effect.succeed(
     Capability.contributes(
       Capabilities.ReactSurface,
       Surface.create({
-        id: 'article',
-        filter: AppSurface.object(AppSurface.Article, Collection.Collection),
-        component: ({ role, data }) => {
-          return <StackArticle attendableId={data.attendableId} role={role} subject={data.subject} />;
-        },
+        id: 'root',
+        filter: AppSurface.object(AppSurface.Article, Template.Data),
+        component: TemplatePanel,
+        props: ({ role, data: { subject, attendableId } }) => ({ role, subject, attendableId }),
       }),
     ),
   ),
