+++ packages/plugins/plugin-terra/src/capabilities/react-surface.ts
@@ -3,27 +3,25 @@
 //
 
 import * as Effect from 'effect/Effect';
-import React from 'react';
 
 import { Capabilities, Capability } from '@dxos/app-framework';
 import { Surface } from '@dxos/app-framework/ui';
 import { AppSurface } from '@dxos/app-toolkit/ui';
 
-import { ScoreArticle } from '#containers';
-import { Score } from '#types';
+import { TerraArticle } from '#containers';
+import { Terra } from '#types';
 
 export default Capability.makeModule(() =>
   Effect.succeed(
     Capability.contributes(Capabilities.ReactSurface, [
       Surface.create({
-        id: 'article',
+        id: 'terra',
         filter: AppSurface.oneOf(
-          AppSurface.object(AppSurface.Article, Score.Score),
-          AppSurface.object(AppSurface.Section, Score.Score),
+          AppSurface.object(AppSurface.Article, Terra.Terra),
+          AppSurface.object(AppSurface.Section, Terra.Terra),
         ),
-        component: ({ data, role }) => (
-          <ScoreArticle role={role} subject={data.subject} attendableId={data.attendableId} />
-        ),
+        component: TerraArticle,
+        props: ({ role, data: { subject, attendableId } }) => ({ role, subject, attendableId }),
       }),
     ]),
   ),
