+++ packages/plugins/plugin-transcription/src/capabilities/react-surface.ts
@@ -3,28 +3,26 @@
 //
 
 import * as Effect from 'effect/Effect';
-import React from 'react';
 
 import { Capabilities, Capability } from '@dxos/app-framework';
 import { Surface } from '@dxos/app-framework/ui';
 import { AppSurface } from '@dxos/app-toolkit/ui';
+import { Transcript } from '@dxos/types';
 
-import { BoardArticle } from '#containers';
-import { Board } from '#types';
+import { TranscriptionArticle } from '#containers';
 
 export default Capability.makeModule(() =>
   Effect.succeed(
     Capability.contributes(Capabilities.ReactSurface, [
       Surface.create({
-        id: 'root',
+        id: 'article.transcript',
         // TODO(wittjosiah): Split into multiple surfaces if this filter proves too strict for non-article roles.
         filter: AppSurface.oneOf(
-          AppSurface.object(AppSurface.Article, Board.Board),
-          AppSurface.object(AppSurface.Section, Board.Board),
+          AppSurface.object(AppSurface.Article, Transcript.Transcript),
+          AppSurface.object(AppSurface.Section, Transcript.Transcript),
         ),
-        component: ({ role, data }) => (
-          <BoardArticle role={role} subject={data.subject} attendableId={data.attendableId} />
-        ),
+        component: TranscriptionArticle,
+        props: ({ role, data: { subject, attendableId } }) => ({ role, subject, attendableId }),
       }),
     ]),
   ),
